feat: Implement Light Server model and enhance Generic Server documentation
Key Changes:
Header Includes: Added Light Server headers to meshx_c_header.h
Code Formatting: Fixed spacing and alignment in Generic model headers
Virtual Destructors: Made destructors explicitly virtual for proper inheritance
Function Fix: Corrected plat_send_msg() call in Light Server - removed invalid data_len parameter
Constants: Made magic numbers constexpr for better type safety
Documentation: Added comprehensive comments for callback registration/deregistration functions
Files Modified:
meshx_c_header.h
 - Added Light Server includes
meshx_base_model_generic.hpp
 - Formatting and virtual destructor fixes
meshx_base_model_light.cpp
 - Fixed platform send function call
meshx_base_model_class.hpp/cpp - Documentation and constant improvements
 call in Light Server - removed invalid data_len parameter
Constants: Made magic numbers constexpr for better type safety
Documentation: Added comprehensive comments for callback registration/deregistration functions
Files Modified:
meshx_c_header.h
 - Added Light Server includes
meshx_base_model_generic.hpp
 - Formatting and virtual destructor fixes
meshx_base_model_light.cpp
 - Fixed platform send function call
meshx_base_model_class.hpp/cpp - Documentation and constant improvements

+++ main/component/meshx/elements/server/src/meshx_cwww_server.c
@@ -471,8 +471,8 @@ static meshx_err_t cwww_prov_control_task_handler(dev_struct_t const *pdev, cont
                 err = meshx_gen_on_off_srv_send_pack_create(
                         CWWW_SRV_EL(rel_el_id).onoff_srv_model->meshx_sig,
                         (uint16_t)el_id,
-                        pdev->meshx_store.net_key_id,
-                        CWWW_SRV_EL(rel_el_id).srv_ctx->app_id,
+                        (uint8_t)pdev->meshx_store.net_key_id,
+                        (uint8_t)CWWW_SRV_EL(rel_el_id).srv_ctx->app_id,
                         CWWW_SRV_EL(rel_el_id).srv_ctx->pub_addr,
                         CWWW_SRV_EL(rel_el_id).srv_ctx->prev_state.on_off,
                         &gen_srv_send
